+++ src/control/console_input.h
@@ -4,7 +4,7 @@
 
 class ConsoleInput : public Input {
 public:
-    auto read_seed() -> std::optional<uint64_t> override;
+    auto read_seed() -> std::optional<uint32_t> override;
     auto read_path() -> std::filesystem::path override;
     auto read_action() -> std::optional<UserAction> override;
 };
